requirements.txt
Django==3.2.10
djangorestframework==3.12.4
drf-yasg==1.20.0
django-cors-headers==3.8.0
djangorestframework_simplejwt==5.3.1